On branch dev Changes to be committed: 	modified:   .gitignore 	renamed:    cgb/cgb.py -> Base/_Base.py 	new file:   Base/_Losses.py 	new file:   Base/__init__.py 	deleted:    cgb/__init__.py 	new file:   model/__init__.py 	new file:   model/cgb.py 	deleted:    test/__init__.py 	deleted:    test/tets_clf.py
