+++ Base/__init__.py
@@ -1,0 +1,7 @@
+from ._Losses import CondensedDeviance, MultiOutputLeastSquaresError
+from ._Base import CondensedGradientBoosting
+
+
+__all__ = ["CondensedGradientBoosting",
+           "CondensedDeviance",
+           "MultiOutputLeastSquaresError"]
