--- cgb/__init__.py
@@ -1,4 +1,0 @@
-from .cgb import cgb_clf
-from .cgb import cgb_reg
-
-__all__ = ["cgb_clf", "cgb_reg"]
