+++ model/__init__.py
@@ -1,0 +1,4 @@
+from .cgb import cgb_clf, cgb_reg
+
+
+__all__ = ["cgb_clf", "cgb_reg"]
